time sync: button sync event, sub-us servo, wrap-safe ts, UART logs
- button (sw0) sync event with red UART log
- fixed-point trimmed-mean clock servo (sub-us offset)
- unsigned 32-bit timestamp mapping (fix negative shared_ts after 2^31)
- auto pulse every 5 ms, periodic log every 1 s
- UART console output (replaces RTT)
- refresh demo PPT (robot + multi-scenario) and one-page intro

+++ include/time_sync.h
@@ -27,13 +27,17 @@ void time_sync_update(uint32_t tx_ts_us, uint32_t local_ts_us);
  *        timebase.
  *
  * Both clocks wrap at the same ~1 MHz rate, so the mapping is a wrap-safe
- * 32-bit addition; do not mix in the local 64-bit GRTC high bits.
+ * unsigned 32-bit addition; do not mix in the local 64-bit GRTC high bits
+ * and do not sign-extend the 32-bit value: after ~35.8 min of uptime the
+ * raw controller time passes 2^31 and a signed interpretation would go
+ * negative.
  *
  * @param local_ts_us Local controller timestamp (32-bit, e.g. the ISO SDU
  *        synchronization reference).
- * @return The corresponding time in the broadcaster's timebase (us).
+ * @return The corresponding time in the broadcaster's timebase (us,
+ *         unsigned 32-bit, same modulus as the TX timestamps).
  */
-int64_t time_sync_to_shared(uint32_t local_ts_us);
+uint32_t time_sync_to_shared(uint32_t local_ts_us);
 
 /**
  * @brief Whether the servo has collected enough samples and is tracking.
